+++ src/ui/event/key/mod.rs
@@ -12,10 +12,10 @@
 #[cfg(test)]
 mod tests;
 
-mod alpha_pad; // KeyAlpha, KeyPad
 mod event; // EventKey, EventKeyFfi, KeyState
 mod key; // Key
 mod media_mods; // KeyMedia, KeyMod, KeyMods
+mod pad; // KeyPad
 
 #[cfg(ffi··)]
 #[cfg_attr(nightly_doc, doc(cfg(ffi··)))]
@@ -24,10 +24,10 @@ mod ffi; // KeyFfi
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
-            alpha_pad::*,
             event::*,
             key::*,
             media_mods::*,
+            pad::*,
         };
         #[cfg(ffi··)]
         pub use super::ffi::*;